Create Language Table and Description
Create a table with 5 rows and 3 columns. Table headers should be S.no., Language, and Used for. Fill the table with the content in exact order:

The row next to the heading row should have 1. as S.no, "HTML" in the Language column, and "Frontend" in the User for column.

The next row should have 2. as S.no, "CSS" in the Language column, and "Frontend" in the User for column.

The next row should have 3. as S.no, "JavaScript" in the Language column, and "Frontend and Backend" in the User for column.

The next row should have 4. as S.no, "Python" in the Language column, and "Backend" in the User for column.


Create a section with class definition-description-list and add heading, of level-3, Definition/ Description list. After heading, use the description list element to describe each language.
*Text for description list and describing the languages have been provided in the boilerplate and commented out.*
*Follow the order of the tags and content to pass the tests*